Add case insensitivity for LIKE queries
Usually, you want your LIKE queries to be case insensitive. Even though it does incur a modest performance hit, it makes the LIKE query much more likely to return proper results.

For example, if you're searching for any records that include "Hello", it's nice to be able to type "Hello", and get results for "hello" as well.